global: clarify exceptions to reloading and freezing, and data -> behavior back
Some community-made tools use techniques such as "freezing" or
"suspending" the game, as to not affect the game while it does not have
focus.

Switch to using "behavior pack" as opposed to "data pack", as it's the
most common term known by runners.

Finally, phrase the beta version restriction to allow for categories
that explicitly allow it, such as 1.16 beta.