global: allow changing the world's difficulty during the run
We are allowing difficulty change, as it positively helps AA and some
niche situations.

Allowing the option itself is not really a big issue compared to the
other settings, and it has been allowed on the Minecraft Java and Legacy
Console boards for a long time now

This change is explicitely only for the difficulty setting only, and no
others.