global: clarify relogging rules for glitchless
Due to Bedrock's many bugs relogging is sometimes easiest, if not the
only way for a run to continue.

To accomodate for these common issues, the rules will have some "wiggle
room", where harmless relogging is allowed, while still banning
relogging with the intent of glitching in an otherwise glitchless run.

Co-Authored-By: Luca Matei Pintilie <luca@lucamatei.com>